// parallel-web/src/types/ResourceMonitor.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export type ResourceMonitor = { cpu_usage_percent: number, memory_usage_percent: number, memory_used_mb: bigint, memory_total_mb: bigint, disk_usage_percent: number, disk_used_gb: number, disk_total_gb: number, };