// parallel-web/src/types/WorkerInfo.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { WorkerCapabilities } from "./WorkerCapabilities";
import type { WorkerStatus } from "./WorkerStatus";

export type WorkerInfo = { id: string, token: string, name: string, status: WorkerStatus, last_heartbeat: string, current_tasks: string[], capabilities: WorkerCapabilities, max_concurrent: number, };